Full stack, Inventory Management system
In this emerging world of technology, every system is computerized and is available online. The Inventory Management of any organization is one of a critical requirement for the business success of the industry. The main objective of this application is to allow the admin, who is usually the Logistics manager of an organization to see the customer orders, check the stock position of the products and raw materials and manage the vendor in a single dashboard. This also allows customers to place their orders online, as well as the Vendor to raise Invoices and collect his payments. The portal believes in providing the most secure and convenient Order Management system, keeping the customers orders confidential.